Update pipeline.m
As mentioned in the meeting, for the RefreshIOBt, I let it call UpdateCalculatorList. Now we just need to implement counting input and output and we'll be good to go.
Also, please test the following features before merging, I am not sure if it is just me or if we need to fix them:
1. Try to select a calculator and click the Edit button; see if it throws an error
2. (probably just me) Try to reload pipelines and see if it would cause a large blank space on top
3. (probably just me) Try to resize the window and see if the buttons are moving accordingly
I will go back to fix those if they are also happening on your computer. But currently, I can't see anything wrong with the code (also, 2 & 3 did not happen when I was using MATLAB R2021a)